move header button to his own file and header
